refactor(frontend): copilot home becomes briefing recap, BriefingCard promoted to organisms
The /copilot empty state no longer duplicates the home surface: the
decisions inbox and team strip live on /home, so CopilotHome renders only
a short briefing recap above the composer.

- Move BriefingCard (+ helpers, tests) to src/components/organisms and give
  it a className prop so callers own its spacing
- Drop NeedsAttentionList and TeamStrip from CopilotHome; delete the
  TeamStrip component (CopilotHome was its only consumer)
- Extract NeedsAttentionRow into its own file, no behaviour change
