+++ autogpt_platform/frontend/src/app/(platform)/copilot/components/CopilotHome/CopilotHome.tsx
@@ -1,20 +1,18 @@
 "use client";
 
 import { ErrorCard } from "@/components/molecules/ErrorCard/ErrorCard";
-import { NeedsAttentionList } from "@/components/organisms/NeedsAttentionList/NeedsAttentionList";
+import { BriefingCard } from "@/components/organisms/BriefingCard/BriefingCard";
 import type { ReactNode } from "react";
-import { BriefingCard } from "./components/BriefingCard/BriefingCard";
-import { TeamStrip } from "./components/TeamStrip/TeamStrip";
 import { useCopilotHome } from "./useCopilotHome";
 
 interface Props {
   fallback: ReactNode;
 }
 
-// The briefing-first block of the copilot home. Mounted *inside*
-// EmptySession rather than beside it, so the experts cohort keeps the
-// onboarding surface (welcome dialog, greeting flow, suggestion themes) and
-// the composer's recipient picker, which live there.
+// The short briefing recap above the composer. It is a recap only — the
+// decisions inbox and team status live on /home — so the copilot's empty
+// state keeps its onboarding surface (welcome dialog, greeting flow,
+// suggestion themes) and the composer's recipient picker.
 export function CopilotHome({ fallback }: Props) {
   const {
     briefing,
@@ -24,29 +22,23 @@ export function CopilotHome({ fallback }: Props) {
     hasBriefing,
   } = useCopilotHome();
 
-  return (
-    <>
-      {/* Briefing card slot — falls back to whatever the caller renders in
-          its place only when a successful fetch says there is no briefing
-          yet; a failed fetch shows the error card so it can't masquerade as
-          "no briefing". Nothing renders until load settles so the fallback
-          doesn't flash for users who do have a briefing. */}
-      {isLoadingBriefing ? null : isBriefingError ? (
-        <ErrorCard
-          context="briefing"
-          httpError={{ message: "Failed to load your briefing" }}
-          onRetry={() => refetchBriefing()}
-          className="mb-5"
-        />
-      ) : hasBriefing && briefing ? (
-        <BriefingCard briefing={briefing} />
-      ) : (
-        fallback
-      )}
+  // Nothing renders until load settles so the fallback doesn't flash for
+  // users who do have a briefing. A failed fetch shows the error card so it
+  // can't masquerade as "no briefing".
+  if (isLoadingBriefing) return null;
 
-      <NeedsAttentionList />
+  if (isBriefingError) {
+    return (
+      <ErrorCard
+        context="briefing"
+        httpError={{ message: "Failed to load your briefing" }}
+        onRetry={() => refetchBriefing()}
+        className="mb-5"
+      />
+    );
+  }
 
-      <TeamStrip />
-    </>
-  );
+  if (!hasBriefing || !briefing) return <>{fallback}</>;
+
+  return <BriefingCard briefing={briefing} className="mb-5" />;
 }
