+++ autogpt_platform/frontend/src/app/(platform)/copilot/components/CopilotHome/__tests__/main.test.tsx
@@ -8,14 +8,9 @@ import {
 } from "@/app/api/__generated__/endpoints/library/library.msw";
 import { getGetV1ListAllExecutionsMockHandler } from "@/app/api/__generated__/endpoints/graphs/graphs.msw";
 import { getGetBriefingsGetLatestBriefingMockHandler200 } from "@/app/api/__generated__/endpoints/briefings/briefings.msw";
-import { getListExpertsMockHandler200 } from "@/app/api/__generated__/endpoints/experts/experts.msw";
-import { getGetV1ListExecutionSchedulesForAUserMockHandler } from "@/app/api/__generated__/endpoints/schedules/schedules.msw";
-import { getGetV2GetPendingReviewsMockHandler200 } from "@/app/api/__generated__/endpoints/executions/executions.msw";
-import type { Expert } from "@/app/api/__generated__/models/expert";
-import type { PendingHumanReviewModel } from "@/app/api/__generated__/models/pendingHumanReviewModel";
 import { EmptySession } from "../../EmptySession/EmptySession";
 
-// Rendered through EmptySession, not CopilotHome directly: the briefing home
+// Rendered through EmptySession, not CopilotHome directly: the briefing recap
 // mounts inside it, and the onboarding surface + composer recipient picker
 // that live there must survive the experts flag being on.
 vi.mock("@/services/feature-flags/use-get-flag", async (importActual) => {
@@ -39,9 +34,7 @@ const baseProps = {
 
 // Guarantees the pulse strip has at least one chip regardless of faker's
 // random defaults — an agent with an external trigger always produces a
-// "listening" sitrep item. Also pins pending reviews to empty since the
-// generated mock otherwise returns 1-10 random reviews by default, which
-// would make the needs-attention slot non-deterministic across tests.
+// "listening" sitrep item.
 function mockPulseStripAgent() {
   const base = getGetV2ListLibraryAgentsResponseMock();
   server.use(
@@ -58,102 +51,6 @@ function mockPulseStripAgent() {
       },
     }),
     getGetV1ListAllExecutionsMockHandler([]),
-    getGetV2GetPendingReviewsMockHandler200([]),
-  );
-}
-
-const pendingReview: PendingHumanReviewModel = {
-  node_exec_id: "ne-1",
-  node_id: "n-1",
-  user_id: "u-1",
-  graph_exec_id: "run-1",
-  graph_id: "g-1",
-  graph_version: 1,
-  payload: { to: "x@y.com" },
-  instructions: "Approve outreach email",
-  editable: true,
-  status: "WAITING",
-  expert_id: "exp-1",
-  expert_name: "Ana",
-  expert_avatar_url: null,
-  agent_name: "Lead Finder",
-  library_agent_id: "lib-1",
-  session_id: null,
-  created_at: new Date(),
-};
-
-const healthyExpert: Expert = {
-  id: "expert-healthy",
-  name: "Sales Scout",
-  avatar_url: null,
-  role: "Sales Researcher",
-  bio: null,
-  skills: [],
-  tagline: "Finds leads while you sleep",
-  identity: "You are a senior sales researcher.",
-  voice_preferences: "",
-  boundaries: "",
-  protected_soul_rules: [],
-  is_template: false,
-  source_template_id: null,
-  is_archived: false,
-  workflows: [],
-  last_run_at: new Date("2026-08-06T10:00:00Z"),
-  last_run_status: "COMPLETED",
-};
-
-const pausedExpert: Expert = {
-  id: "expert-paused",
-  name: "Support Bot",
-  avatar_url: null,
-  role: "Support Triager",
-  bio: null,
-  skills: [],
-  tagline: "Triages tickets",
-  identity: "You are a support triager.",
-  voice_preferences: "",
-  boundaries: "",
-  protected_soul_rules: [],
-  is_template: false,
-  source_template_id: null,
-  is_archived: false,
-  workflows: [],
-  schedules_paused_at: new Date("2026-08-05T10:00:00Z"),
-};
-
-const needsSetupExpert: Expert = {
-  id: "expert-needs-setup",
-  name: "Ops Analyst",
-  avatar_url: null,
-  role: "Operations Analyst",
-  bio: null,
-  skills: [],
-  tagline: "Keeps the ops dashboard current",
-  identity: "You are an operations analyst.",
-  voice_preferences: "",
-  boundaries: "",
-  protected_soul_rules: [],
-  is_template: false,
-  source_template_id: null,
-  is_archived: false,
-  workflows: [
-    {
-      id: "wf-1",
-      store_listing_version_id: "slv-1",
-      library_agent_id: "lib-1",
-      graph_id: "graph-1",
-      name: "Daily Report",
-      description: null,
-      schedule_cron: "0 9 * * *",
-      schedule_id: null,
-    },
-  ],
-};
-
-function mockTeamStrip() {
-  server.use(
-    getListExpertsMockHandler200([healthyExpert, pausedExpert]),
-    getGetV1ListExecutionSchedulesForAUserMockHandler([]),
   );
 }
 
@@ -232,68 +129,35 @@ test("renders briefing sections when a briefing is available", async () => {
   const finding = screen.getByText(/Found 3 leads/);
   expect(finding.textContent).toContain("Lead Finder");
 
-  // The card does not repeat the decisions: the needs-attention list below
+  // The recap does not repeat the decisions: the needs-you list on /home
   // shows the same pending reviews, and it can act on them.
   expect(screen.queryByText(/Needs your decision/)).toBeNull();
 });
 
-test("renders the team strip with hired experts", async () => {
+test("keeps the decisions inbox and team status off the copilot recap", async () => {
+  // Both moved to /home. Proving the pending-reviews query never fires is
+  // what keeps the inbox from creeping back in: absent text alone would
+  // also pass while the request was still in flight. (The experts query has
+  // no such tell — the composer's recipient picker still needs it.)
   mockPulseStripAgent();
   server.use(getGetBriefingsGetLatestBriefingMockHandler200(null));
-  mockTeamStrip();
-  render(<EmptySession {...baseProps} />);
 
-  expect(await screen.findByText("Sales Scout")).toBeDefined();
-  expect(screen.getByText("Support Bot")).toBeDefined();
-  expect(screen.getByText("Paused")).toBeDefined();
+  const requestedPaths: string[] = [];
+  function record({ request }: { request: Request }) {
+    requestedPaths.push(new URL(request.url).pathname);
+  }
+  server.events.on("request:start", record);
 
-  // Each Chat link names its expert, so screen readers can tell the
-  // repeated links apart.
-  expect(
-    screen
-      .getByRole("link", { name: "Chat with Sales Scout" })
-      .getAttribute("href"),
-  ).toBe("/copilot?expertId=expert-healthy");
-  expect(
-    screen
-      .getByRole("link", { name: "Chat with Support Bot" })
-      .getAttribute("href"),
-  ).toBe("/copilot?expertId=expert-paused");
-});
+  try {
+    render(<EmptySession {...baseProps} />);
+    await screen.findByText("What's happening with your agents");
 
-test("uses singular grammar for a single workflow needing setup", async () => {
-  mockPulseStripAgent();
-  server.use(
-    getGetBriefingsGetLatestBriefingMockHandler200(null),
-    getListExpertsMockHandler200([needsSetupExpert]),
-    getGetV1ListExecutionSchedulesForAUserMockHandler([]),
-  );
-  render(<EmptySession {...baseProps} />);
-
-  expect(await screen.findByText("1 workflow needs setup")).toBeDefined();
-});
-
-test("renders a needs-attention row when a review is pending", async () => {
-  mockPulseStripAgent();
-  server.use(
-    getGetBriefingsGetLatestBriefingMockHandler200(null),
-    getGetV2GetPendingReviewsMockHandler200([pendingReview]),
-  );
-  render(<EmptySession {...baseProps} />);
-
-  expect(await screen.findByText("Approve outreach email")).toBeDefined();
-});
-
-test("does not render the needs-attention slot when there are no pending reviews", async () => {
-  mockPulseStripAgent();
-  server.use(
-    getGetBriefingsGetLatestBriefingMockHandler200(null),
-    getGetV2GetPendingReviewsMockHandler200([]),
-  );
-  render(<EmptySession {...baseProps} />);
-
-  await screen.findByText("What's happening with your agents");
-  expect(screen.queryByText(/Needs your attention/)).toBeNull();
+    expect(requestedPaths.some((path) => path.includes("review"))).toBe(false);
+    expect(screen.queryByText(/Needs your attention/)).toBeNull();
+    expect(screen.queryByRole("link", { name: /Chat with/ })).toBeNull();
+  } finally {
+    server.events.removeListener("request:start", record);
+  }
 });
 
 test("shows an error card instead of the pulse strip when the briefing fetch fails", async () => {
@@ -309,25 +173,9 @@ test("shows an error card instead of the pulse strip when the briefing fetch fai
   expect(screen.queryByText("What's happening with your agents")).toBeNull();
 });
 
-test("shows an error card when the pending-reviews fetch fails", async () => {
-  mockPulseStripAgent();
-  server.use(
-    getGetBriefingsGetLatestBriefingMockHandler200(null),
-    http.get("/api/proxy/api/review/pending", () =>
-      HttpResponse.json({ detail: "boom" }, { status: 500 }),
-    ),
-  );
-  render(<EmptySession {...baseProps} />);
-
-  expect(
-    await screen.findByText("Failed to load pending reviews"),
-  ).toBeDefined();
-  expect(screen.queryByText(/Needs your attention/)).toBeNull();
-});
-
 test("keeps the onboarding surface and recipient picker with the experts flag on", async () => {
   // hire-experts and onboarding-brain-dump are independent flags aimed at
-  // overlapping beta cohorts; the briefing home must not cancel the other
+  // overlapping beta cohorts; the briefing recap must not cancel the other
   // rollout out from under a brand-new user.
   mockPulseStripAgent();
   server.use(getGetBriefingsGetLatestBriefingMockHandler200(null));
