+++ autogpt_platform/frontend/src/components/organisms/BriefingCard/helpers.test.ts
